Automated g4 rollback of changelist 205815425.
Please note that this was created with g4 rollback, but the version number in the Wycheproof file regresses. I'm unsure whether that's desired. The BoringSSL change works for both the old and new files so can be landed without touching Wycheproof if that's better.

*** Reason for rollback ***

Fix to BoringSSL scripts now included.

*** Original change description ***

Adding the curve back to the test cases to avoid failing BoringSSL tests.

***

PiperOrigin-RevId: 205957468
GitOrigin-RevId: c0afaa191d5509050bbea307265b07bca61b4714