Updating the test vectors for ECDH.
The main reason is this:
Version 1.68 of Bouncycastle no longer checks the cofactor of public keys.
This does not change the outcome of ECDH computations since the cofactor is
not used there.
Wycheproof tests ECDH computations with public keys with modified parameters.
The test expects that the ECDH computation always uses the parameters of the
private key. Otherwise the test fails.
In some cases Wycheproof requires that mismatched or invalid parameters are
always rejected. I.e., in the previous version a cofactor larger than 2^32
would always be an invalid test vector, since this is a limit specified by
FIPS-PUB 186-4.
This restriction has been relaxed. There is now simply a flag for such keys.
Hence the tests now pass when an implementation does not use the cofactor
and simply ignores it.

Other changes:
- adding more test vectors for ASN parsing.
- adding a few more flags.

NOKEYCHECK=True
PiperOrigin-RevId: 375040843