Refactoring JsonSignatureTest:
The main goal of this CL is to refactor the Wycheproof tests with test vectors
so that the tests can be called more easily from a third party.

To achieve this goal two classes TestVectors and TestResult have been added.
TestVectors contains the inputs of the test. It should allow running individual
tests. TestResult contains the output of the test.
The tests itself have been split into methods that perform the test and
methods that print the results.

One issue is distinguishing between a test valid vector that failed because
there is an implementation error or a test vector that failed because the
primitive is not supported:
The current implementation is to check if any test valid test vector passed.
If none passed then the primitive is probably not implemented or properly supported
and hence the test is skipped.

Things explicitly not done in this CL:
* The formatting of the test results is poor.
  I expect this to change this when the usage become more clear.
* It may be possible to give some preliminary assessment of the failures.
* Additional parameters may be added to TestVectors to parameterize the tests.

NOKEYCHECK=True
PiperOrigin-RevId: 457668413