Modifying TestUtil for an upcoming version of jdk.
NOTE: This CL is not a solution for b/238938657. Rather it fixes some other
problems, so that tests against the latest jdk version are possible. Currently blaze test --config=java_runtime_head OpenJDKTest fails,
because some providers are not longer public.

The old jdk version has 19 failed, 196 passed, and 58 skipped tests.
The new jdk version has  3 failed, 192 passed, and 78 skipped tests.

The difference is due to the removal of a number of elliptic curves and a number of bug fixes. The curves that were removed are not frequently used, and their implementations contain some bugs.

There are a number of bugs that have been solved:
238938657: This CL tests the new implementations of ECDSA and ECDH.
           No problems are found. Some existing issues with ECDSA and ECDH
           disappear.
237634445: Obsolete since secp224r1 is no longer supported.
238074987: Also obsolete since secp224r1 is no longer supported.
237635484: Will disappear since secp256k1 is no longer supported.
238722814: Oracle fixed the encoding of XDH keys. The tests now pass.
237639284: Fixed in the new version.
237638289: Fixed in the new version.

Keymaster contains some patches that could be removed:
e.g., the parser for DER encoded signatures has been improved and the following
patch is no longer necessary.
https://source.corp.google.com/piper///depot/google3/java/com/google/security/keymaster/KeymasterECDSAPublicKey.java;rcl=450642578;l=167

NOKEYCHECK=True
PiperOrigin-RevId: 461599370