Refactoring CipherOutputStreamTest.java * The tests are now skipped if the cipher is not supported. * Tests are now split into testing encryption, decryption and decryption of   corrupt ciphertexts.
CipherOutputStream has the unfortunate behaviour that a decryption failure
because of an incorrect tag results in an empty plaintext rather than an
exception.
Oracle isn't changing this:
http://www.oracle.com/technetwork/java/javase/8u171-relnotes-4308888.html

The main purpose of the test is to ensure that a provider does not behave
even worse by returning partial plaintext.

Before the CL tests pass if they return empty plaintext. With this CL
tests only pass if there is an exception and either fail or are skipped
if they return empty plaintext.

The corruption tests pass under jdk11 they are skipped under jdk19.

NOKEYCHECK=True
PiperOrigin-RevId: 468142569