Adding more tests for ECDH.
Some provider behave differently when the curve is specified by using

  AlgorithmParameters parameters = AlgorithmParameters.getInstance("EC");
  parameters.init(new ECGenParameterSpec(name));
  ECParameterSpec spec = parameters.getParameterSpec(ECParameterSpec.class);

than when generating the parameters with

  new ECParameterSpec(curveSpec, g, n, h);

A particular curious case is Conscrypt, which does seem to perform an ECDH computation when the second format is used to specify brainpool curves.
Such a behaviour would be an unwanted feature. Conscrypt does not officially support brainpool curves.
(To verify that Conscrypt does indeed compute the ECDH, I'v modified
the result compute_key in OpenSSLECDHKeyAgreement and checked that the test
receives the modified result).

On the other hand jdk11 fails to do an ECDH computation when brainpoolP224 and
the second format is used.

BouncyCastle accepts all curves and all formats.

NOKEYCHECK=True
PiperOrigin-RevId: 474504607