Updating the test vectors for XDH.
The OpenJDK provider generated incorretly formatted private keys.
https://bugs.openjdk.org/browse/JDK-8213363
BouncyCastle now seems to have added fallback code that accepts these keys.
As a result the test vectors with the invalid encoding are changed
to have status LEGACY, which means that a provider may or may not
accept such keys. If they accept the key then the tests expect a
correct computation.

NOKEYCHECK=True
PiperOrigin-RevId: 482762210