Updating CBC/PKCS5 test vectors.
The update removes a test failure for BouncyCastleTest:
When trying to unpad a decrypted message then BouncyCastle throws an InvalidBlockSizeException if the padded message is empty and a BadPaddingException for other invalid padding. Typically one does not want distinguishable exceptions,
since this simplifies chosen ciphertext attacks. In this case however the
exception does not leak additional information.

The update removes the BadPadding flag from empty ciphertexts.
This has the effect that the test comparing exceptions no longer fails,
since the test for exceptions only requires that decrypting ciphertexts
with BadPadding flags lead to indistinguishable exceptions.

NOKEYCHECK=True
PiperOrigin-RevId: 489972463