Accepting IllegalStateException in Xdh. KeyAgreement.generateSecret only throws ShortBufferException and IllegalStateException. BouncyCastle throws IllegalStateException for example when the shared secret is all zero.
Allowing some invalid keys:
OpenJDK used an incorrect encoding of X25519 and X448 keys.
BouncyCastle has added fallback code to accept such keys.
Test vectors where the public keys use this encoding have a "LEGACY" flag.
The tests no pass, if XDH is computed with such test vectors and if
the result is the same as the result with the correct key.

NOKEYCHECK=True
PiperOrigin-RevId: 490732496