Cleaning up exceptions.
The goal here is that allTests should catch all checked exceptions
and any unchecked exceptions thrown by the provider.
If allTests fails then this should be an omission in the test setup.
OSS-Fuzz wants to generate automatic bug reports. To avoid false
positives it is necessary to better distinguish between actually failing
tests and tests that are incorrectly set up.

NOKEYCHECK=True
PiperOrigin-RevId: 490888676