Adding expected values to ECDH exchanges with modified keys.
This is needed for testing ECDH with both OpenJDK and Conscrypt installed.
If the SUNEC provider is installed before Conscrypt then the following can
happen if the public key uses an incorrect curve specification, with a valid
point on the curve of the private key:

SUNEC accepts the invalid public key in KeyFactory.
SUNEC does not perform KeyAgreement.generateSecret.
JCE searches for another provider and uses Conscrypt.
Conscrypt seems to assume that the curve has already (or simply checks the
public point only) and computes the ECDH exchange on the curve of the private
key.

NOKEYCHECK=True
PiperOrigin-RevId: 491338026