Updating SecureRandomTest.java
This CL adds test for new SecureRandom(seed);
The result is that the behavior depends a lot on which providers are installed.
E.g., if no provider with SecureRandom instances is installed then java uses
SHA1PRNG as default and the result is deterministic.
If for example the SUN provider is installed and jdk11 is being used then
NativePRNG is being used and the result is non-deterministic.

NOKEYCHECK=True
PiperOrigin-RevId: 495312129