Documenting exceptions thrown by the tests.
For OSS Fuzz it is important to distinguish between broken providers and
broken test setups. Hence it is important to document the exceptions thrown
by the tests. Generally the following should hold:

The methods allTests all return an instance of type TestResult.
This instance contains information about the test failures and can be used
to determine the severity of the failed tests.
When these methods throw an Exception the the test itself is likely
misconfigured.

Tests that take a filename with test vectors as input typically throw the
following exceptions:

AssumptionViolatedException when the test was skipped. This should indicate
    that the primitive is not supported.
AssertionError when the test failed. This should indicate a bug in the provider
IOException when the test vectors could not be read. This indicates an
    incorrect test setup.
other RuntimeExceptions. This is an unexpected exception that should typically
    not happen.

NOKEYCHECK=True
PiperOrigin-RevId: 496888808