Refactoring jose4j/JsonWebEncryptionTest.java.
Adding the expected plaintext to jwe tests for jose4j.
Distinguishing between expected exceptions, and unexpected ones.
Logging the stack trace only when the test fails.
Adding a list of flaky tests.

This CL does not fix the cause of b/261217218. It just suppresses it and
gives better logs.

NOKEYCHECK=True
PiperOrigin-RevId: 498579465