Splitting the schemas for JWE, JWK and JWS into three separate files.
This allows a better verification of the format of the test vectors.
Tests can always assume that the test vectors have correct format.
Only the fields that are being tested (i.e., "jwe" and "jws")
can be malformed.
google3/third_party/wycheproof/py/test/schema_test.py checks that
the test vectors are compatible with the JSON schemas.

NOKEYCHECK=True
PiperOrigin-RevId: 499765166