Modifying the JWE test vectors for jwcrypto.
Adding more test vectors with 256-bit symmetric algorithms.
Adding a field "enc" to the test vectors. This is the same value that
is encoded in the first part of "jwe" (or the value that was used for
the symmetric encryption if no value is encoded in the "jwe" part.)

The motivation for this change is that coverage of jwcrypto with test
vectors is currently difficult. jwcrypto restricts the content encryption algorithms. Hence valid test vectors may fail because of an unknown/unsupported
algorithm. The new field allows to determine if restricted algorithm are
properly rejected.

NOKEYCHECK=True
PiperOrigin-RevId: 502562034