Patching RSA PKCS #1 padding oracles in jose4j. This is a proposal for the case that we want to patch the vulnerability first before reporting it.
To avoid timing attacks stronger checks are done when using RSA1_5.
Before this CL a sender could wrap for example a 128-bit AES key, with RSA1_5
but claim in the jwe header that A256-GCM was used. Even though the key size
and the algorithm do not match a ciphertext would still decrypt.
With this CL decryption will fail.

There are additional cases where jose4j does not verify if the required
algorithm matches the actual algorithm (see b/264522425). Such cases
are not addressed in this CL.

Presubmit train is here:
https://test.corp.google.com/OCL:504237156:BASE:504224147:1674565845974:b5e3d061

NOKEYCHECK=True
PiperOrigin-RevId: 505077225