Adding test vectors for malformed base64 encodings.
RFC 7515 specifies in Section 2 additional
characters such as spaces and line breaks are not
allowed.

Base64 decoders are frequently not strict about their inputs.
For example base64 in python explicitely allows white space and ignores it.
In other cases the decoders use lookuptables in such a way that invalid characters have a value of 0 or -1.

The new test vectors contain invalid encoded Hmacs. They have been constructed
so that they may fail if base64 decoding is not strict.

NOKEYCHECK=True
PiperOrigin-RevId: 509757096