Adding basic test vectors for PBES2. This is based on RFC 2898.
The files are incompletes since I don't have code for test vector generation
yet. PBES2 generates a random key from a password and salt. Then uses
this key and additional parameters (in this case an IV) to encrypt.

RFC 2898 only defines PBES2 with unauthenticated encryption.
In particular it uses CBC mode with PKCS #5 padding.

The test vectors have flags depending on the password:
"printable": The password consists of printable ASCII characters.
             These are test vectors that can be verified by the SUN provider.
"Ascii": The password consists of ASCII characters.
"Utf8": The password is a valid UTF-8 encoding. These are passwords for which
        the SUN provider can compute the underlying PBKDF function correctly.
"NonUtf8": passwords that are not valid UTF-8 encodings. PBES2 is defined for
        these values, but it is not clear yet if there is a provider that
        supports these passwords.

NOKEYCHECK=True
PiperOrigin-RevId: 515365503