Remove test harnesses, defunct tooling (#127)
* remove java test harnesses, associated tooling

Upstream library authors are best positioned to maintain this sort of
code, leaving the community managed Wycheproof contributors to focus on
JSON test vectors.

* remove javascript test harnesses, associated tooling

For the same reasons articulated in the previous commit.

* remove kokoro CI bits

This relied on the now-removed Java harnesses, and was specific to an
internal Google CI system that's not in use now that the repo has
transitioned to community management.