convert all v1 vectors to use common notes schema
Reference the common notes schema from all v1 schemas.

Some v1 vectors were sharing a schema with v0, however the notes field
structure is not compatible between the two. This worked previously
because the schema definition didn't cover the individual note objects.

Now that we've tightened the definition, we need to split those shared
schemas. Soon we will delete the v0 vectors and associated schema files,
so the duplication is temporary.