Add ECDSA tests with non-minimally encoded DER tags (#161)
DER tags are typically one byte, but when the tag number exceeds 30,
they are encoded in a multi-byte tag form. The tag number bits of the
first byte is set to all ones (i.e. 31) and then the actual tag number
is encoded in the big-endian base128 encoding. (See X.690 (02/2021)
8.1.2.4.)

This encoding is only for larger tag numbers, so these should be
rejected by an ECDSA signature verifier. This case was missing from the
original Wycheproof tests.

Although this is a general DER issue, it makes sense to test this
specifically in ECDSA test vectors because ECDSA is deep in the crypto
layer of libraries, and so it is not uncommon for this to be implemented
with slightly different parsers than the rest of the system.

These tests are tagged InvalidEncoding and not BerEncodedSignature
because non-minimal integers is forbidden in BER too. X.690 (02/2021)
8.1.2 says that, for even BER, the single-octet encoding is the encoding
for tags 0 to 30 and the multi-octet encoding is only for 31 and up.

I've added the tests to both testvectors and testvectors_v1 since it was
easy, though hopefully that situation will settle out soon.