remove testvectors directory
The project is consolidating maintenance effort on `testvectors_v1/`, so
this commit removes the original "v0" `testvectors/` directory.

Users relying on the removed content can continue to use Wycheproof
the git tag created from the parent of this commit
(`wycheproof-v0-vectors`), but should consider adopting the v1 vectors
in order to receive new algorithm coverage and other related
fixes/updates.