Provide JSON schema for EcCurveTest
The schema was mostly modeled after the one for the ML-KEM tests. It
marks the generatorVersion as deprecated like in all other schemas
and adds a source object to the TestGroup. Remove it from the list
of missing schemas in the readme and the linter.

ec_prime_order_curve_test.json limits itself in scope to curves of
prime order over prime fields. This doesn't seem entirely correct:
First, its format is only really suitable for the short Weierstrass
representation of a curve over a finite field. Second, that the order
is prime is a desirable common feature of the curves it currently
contains rather than a necessity. Perhaps the test should be renamed
and then it could easily be extended in scope.